requirements.txt
einops
pillow
pandas
imageio
opencv-python
scipy
timm
tqdm
torchmetrics
yacs
pytorch_wavelets
pywavelets
matplotlib